+++ branches/igt_cuda_merge/Modules/RegistrationCuda/ntkElasticTransformation.cpp
@@ -26,8 +26,8 @@ ntk3DData* ntkElasticTransformation::applyTransformation(ntkDeformationSpline* s
     tempPow=1;
     tempSize=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
-    tempSize=(int)pow(2, -splineSizeLevel);
+    tempPow=pow((double)2, (double)splineSizeLevel);
+    tempSize=(int)pow((double)2, (double)-splineSizeLevel);
   }
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
 
@@ -101,8 +101,8 @@ ntkFloatDimension ntkElasticTransformation::getNewPositionFromOld(ntkDeformation
     tempPow=1;
     tempSize=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
-    tempSize=(int)pow(2, -splineSizeLevel);
+    tempPow=pow((double)2, (double)splineSizeLevel);
+    tempSize=(int)pow((double)2, (double)-splineSizeLevel);
   }
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
 
@@ -147,8 +147,8 @@ ntk3DData* ntkElasticTransformation::getDeformationFieldImage(ntkDeformationSpli
     tempPow=1;
     tempSize=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
-    tempSize=(int)pow(2, -splineSizeLevel);
+    tempPow=pow((double)2, (double)splineSizeLevel);
+    tempSize=(int)pow((double)2, (double)-splineSizeLevel);
   }
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
 
@@ -258,6 +258,7 @@ ntkDeformationSpline *ntkElasticTransformation::getReverseTransformationSpline(n
   ntkTensor* newZTensor=new ntkTensor(newZ, splineSize);
   free(newZ);
   
+  /*
   Matrix matTx(splineSize.x,splineSize.x,0.0);
   Matrix matTxI;
   Matrix matTy(splineSize.y,splineSize.y,0.0);
@@ -270,7 +271,7 @@ ntkDeformationSpline *ntkElasticTransformation::getReverseTransformationSpline(n
       matTx(i,j)=func->getValue(0);
     }
   }
-
+*/
   delete func;
   delete newXTensor;
   delete newYTensor;
